+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Fixed `nf-core modules install` not working when installing from branch with `-b` ([#1218](https://github.com/nf-core/tools/issues/1218))
 * Added prompt to choose between updating all modules or named module in  `nf-core modules update`
 * Check if modules is installed before trying to update in `nf-core modules update`
+* Verify that a commit SHA provided with `--sha` exists for `install/update` commands
 
 ## [v2.0.1 - Palladium Platypus Junior](https://github.com/nf-core/tools/releases/tag/2.0.1) - [2021-07-13]